+++ tests/test_pyenv.py
@@ -17,7 +17,7 @@ def mk_python(basename, prefix=None, base_prefix=None, executable=True, content=
         folder = runez.to_path(".pyenv/versions") / basename / "bin"
 
     else:
-        version = Version.from_text(os.path.basename(basename))
+        version = Version(os.path.basename(basename))
         if basename.startswith("./"):
             folder = runez.to_path(os.path.dirname(basename[2:]))
 
@@ -46,13 +46,14 @@ def test_depot(temp_folder, logged):
     mk_python("8.6.1")
     mk_python("8.7.2")
     mk_python("miniforge3-22.11.1-4/9.11.2")
+    mk_python("pypy-9.8.7/9.8.7")
     mk_python("8.5.4", content=["invalid"])
     mk_python("8.5.5", content=["invalid-version", "prefix", "prefix"])
     runez.symlink(".pyenv/versions/8.6.1", ".pyenv/versions/8.6", logger=None)
     runez.symlink(".pyenv/versions/8.6.1", ".pyenv/versions/python8.6", logger=None)
 
     depot = PythonDepot(".pyenv/versions/**")
-    assert len(depot.available_pythons) == 3
+    assert len(depot.available_pythons) == 4
 
     assert str(depot.find_python("8.5.4")) == "8.5.4 [not available]"
     invalid = depot.find_python(".pyenv/versions/8.5.4")
@@ -61,7 +62,7 @@ def test_depot(temp_folder, logged):
     assert str(bad_version) == ".pyenv/versions/8.5.5 [invalid version 'invalid-version']"
 
     versions = [p.mm_spec.canonical for p in depot.available_pythons]
-    assert versions == ["cpython:8.7", "cpython:8.6", "conda:9.11"]
+    assert versions == ["pypy:9.8", "cpython:8.7", "cpython:8.6", "conda:9.11"]
 
     # Verify that latest available is found (when under-specified)
     p8 = depot.find_python("8")
@@ -415,8 +416,9 @@ def test_venv_from_project():
     ]
 )
 def test_pep_sample(given_version, expected):
-    version = Version.from_text(given_version, strict=False)
+    version = Version(given_version)
     assert version.is_valid
+    assert version.ignored is None
     assert str(version) == given_version
     actual = version.components
     if version.prerelease:
@@ -426,19 +428,19 @@ def test_pep_sample(given_version, expected):
 
 
 def test_version():
-    loose = Version("v1.0.dirty", strict=False)
+    loose = Version("v1.0.dirty", canonical=None)
     assert loose.is_valid
     assert loose.ignored == ".dirty"
     assert str(loose) == "1.0"
     assert loose.pep_440 == "1.0"
 
-    invalid = Version("v1.0.dirty", strict=True)
+    invalid = Version("v1.0.dirty")
     assert not invalid.is_valid
     assert str(invalid) == "v1.0.dirty"
     assert invalid.ignored == ".dirty"
     assert loose > invalid
 
-    dev101 = Version.from_text("0.0.1dev101")
+    dev101 = Version("0.0.1dev101")
     assert not dev101.is_final
     assert dev101.is_valid
     assert not dev101.is_dirty
@@ -503,11 +505,11 @@ def test_version():
     assert v1 > []
     assert v1 > [5, 2, 3, 4, 5, 6]
 
-    v1foo = Version.from_text("1foo")  # Ignore additional text
+    v1foo = Version("1foo", canonical=None)  # Ignore additional text
     assert v1 == v1foo
 
-    vrc = Version.from_text("1.0rc4-foo")
-    vdev = Version.from_text("1.0a4.dev5-foo")
+    vrc = Version("1.0rc4-bar", canonical=None)
+    vdev = Version("1.0a4.dev5-foo", canonical=None)
     assert vdev.pep_440 == "1.0a4.dev5"
     assert vdev.ignored == "-foo"
     assert vrc.is_valid
@@ -541,16 +543,15 @@ def test_version():
 
 
 def test_version_extraction():
-    x = Version.from_text("foo, version 1.0a4.dev5\nbar baz")
+    x = Version.extracted_from_text("foo, version 1.0a4.dev5\nbar baz")
     assert str(x) == "1.0a4.dev5"
-    assert x.ignored == "foo, version"
 
-    # .canonical() can be used to filter out invalid versions as None
-    assert Version.from_text("Python 3.8.6", strict=True) is None
-    p38 = Version.from_text("Python 3.8.6")
+    p38 = Version("Python 3.8.6")
+    assert not p38.is_valid
+
+    p38 = Version.extracted_from_text("Python 3.8.6")
     assert str(p38) == "3.8.6"
     assert p38.is_valid
-    assert p38.ignored == "Python"
 
 
 def test_version_comparison():
@@ -572,7 +573,7 @@ def test_version_comparison():
     assert v == "2.11.0 2.11.1.dev1 2.11.1.dev2+b 2.11.1.dev11+a.dirty 2.11.1 1!1.0"
 
     v = "1.2rc1 1.2rc2.dev05 1.2rc2.dev4 1.2a1.dev1"
-    v = runez.flattened(v, split=" ", transform=Version.from_text)
+    v = runez.flattened(v, split=" ", transform=Version)
     assert all(x.is_valid for x in v)
     v = sorted(v)
     v = runez.joined(v)
@@ -700,33 +701,38 @@ def test_version_ordering():
 def verify_ordering(expected):
     # Jumble the given list of versions a bit, then sort them and verify they sort back to 'expected'
     given = sorted(expected, key=lambda x: x.text)
-    x = given.pop(len(given) // 2)
-    given.append(x)
+    mid_item = given.pop(len(given) // 2)
+    given.append(mid_item)
 
     assert given != expected
     assert sorted(given) == expected
 
 
 def test_version_pep_440():
-    vpost = Version.from_text("1.2.post")
+    vpost = Version("1.2.post")
     assert vpost.is_valid
     assert not vpost.prerelease
     assert str(vpost) == "1.2.post"
     assert vpost.pep_440 == "1.2.post0"
 
-    vrev5 = Version.from_text("1.2rev05")
+    vrev5 = Version("1.2rev05")
     assert str(vrev5) == "1.2rev05"
     assert vrev5.is_valid
     assert vrev5.pep_440 == "1.2.post5"
     assert vrev5 > vpost
 
-    vr6 = Version.from_text("1.2r6")
+    vrev5_canonical = Version("1.2rev05", canonical=True)
+    assert str(vrev5_canonical) == "1.2.post5"
+    assert vrev5_canonical.pep_440 == vrev5_canonical.text
+    assert vrev5_canonical == vrev5
+
+    vr6 = Version("1.2r6")
     assert str(vr6) == "1.2r6"
     assert vr6.is_valid
     assert vr6.pep_440 == "1.2.post6"
     assert vr6 > vrev5
 
-    vrev5dev3 = Version.from_text("1.2rev05.dev3")
+    vrev5dev3 = Version("1.2rev05.dev3")
     assert vrev5dev3.prerelease
     assert str(vrev5dev3) == "1.2rev05.dev3"
     assert vrev5dev3.pep_440 == "1.2.rev5.dev3.post5"
@@ -734,6 +740,6 @@ def test_version_pep_440():
     assert vrev5 > vrev5dev3
     assert vr6 > vrev5dev3
 
-    vrc1 = Version.from_text("v1.39.4-rc.1")
+    vrc1 = Version("v1.39.4-rc.1")
     assert str(vrc1) == "1.39.4-rc.1"
     assert vrc1.pep_440 == "1.39.4rc1"